requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy[asyncio]==2.0.35
alembic==1.13.3
asyncpg==0.29.0
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
httpx==0.27.2

# Observability - Logging
structlog==24.4.0
python-json-logger==2.0.7

# Observability - Metrics
prometheus-client==0.21.0

# Observability - Tracing (OpenTelemetry)
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-instrumentation-sqlalchemy==0.48b0
opentelemetry-instrumentation-asyncpg==0.48b0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-prometheus==0.48b0

# Payment Processing
stripe==11.1.1

# Admin Authentication & Security
argon2-cffi==23.1.0
pyjwt==2.9.0
pyotp==2.9.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
